remove resource_helper.h/cpp & move functions to utility header with other conversion functions
